Revert "[Test Framework] Support dynamic echo workloads (#30675)"
This reverts commit ab6cc48134a698d7ad218a83390fe27e8098919f.
