+++ pkg/test/framework/components/echo/kube/workload.go
@@ -17,7 +17,6 @@ package kube
 import (
 	"context"
 	"fmt"
-	"sync"
 	"time"
 
 	"github.com/hashicorp/go-multierror"
@@ -40,92 +39,92 @@ const (
 
 var _ echo.Workload = &workload{}
 
-type workloadConfig struct {
-	pod        kubeCore.Pod
-	hasSidecar bool
-	grpcPort   uint16
-	cluster    cluster.Cluster
-	tls        *common.TLSSettings
-}
-
 type workload struct {
 	*client.Instance
 
-	workloadConfig
+	pod       kubeCore.Pod
 	forwarder istioKube.PortForwarder
 	sidecar   *sidecar
+	cluster   cluster.Cluster
 	ctx       resource.Context
-	mutex     sync.Mutex
 }
 
-func newWorkload(cfg workloadConfig, ctx resource.Context) (*workload, error) {
-	w := &workload{
-		workloadConfig: cfg,
-		ctx:            ctx,
-	}
-
-	// If the pod is ready, connect.
-	if err := w.Update(cfg.pod); err != nil {
+func newWorkload(pod kubeCore.Pod, sidecared bool, grpcPort uint16, cluster cluster.Cluster,
+	tls *common.TLSSettings, ctx resource.Context) (*workload, error) {
+	// Create a forwarder to the command port of the app.
+	var forwarder istioKube.PortForwarder
+	if err := retry.UntilSuccess(func() error {
+		fw, err := cluster.NewPortForwarder(pod.Name, pod.Namespace, "", 0, int(grpcPort))
+		if err != nil {
+			return fmt.Errorf("new port forwarder: %v", err)
+		}
+		if err = fw.Start(); err != nil {
+			fw.Close()
+			return fmt.Errorf("forwarder start: %v", err)
+		}
+		forwarder = fw
+		return nil
+	}, retry.Delay(1*time.Second), retry.Timeout(10*time.Second)); err != nil {
 		return nil, err
 	}
 
-	return w, nil
-}
-
-func (w *workload) IsReady() bool {
-	w.mutex.Lock()
-	connected := w.isConnected()
-	w.mutex.Unlock()
-	return connected
-}
-
-func (w *workload) Update(pod kubeCore.Pod) error {
-	w.mutex.Lock()
-	defer w.mutex.Unlock()
-
-	if isPodReady(pod) && !w.isConnected() {
-		if err := w.connect(pod); err != nil {
-			return err
-		}
-	} else if !isPodReady(pod) && w.isConnected() {
-		w.pod = pod
-		return w.disconnect()
+	// Create a gRPC client to this workload.
+	c, err := client.New(forwarder.Address(), tls)
+	if err != nil {
+		forwarder.Close()
+		return nil, fmt.Errorf("grpc client: %v", err)
 	}
 
-	// Update the pod.
-	w.pod = pod
-	return nil
+	var s *sidecar
+	if sidecared {
+		if s, err = newSidecar(pod, cluster); err != nil {
+			return nil, err
+		}
+	}
+
+	return &workload{
+		pod:       pod,
+		forwarder: forwarder,
+		Instance:  c,
+		sidecar:   s,
+		cluster:   cluster,
+		ctx:       ctx,
+	}, nil
 }
 
 func (w *workload) Close() (err error) {
-	w.mutex.Lock()
-	defer w.mutex.Unlock()
-
-	if w.isConnected() {
-		return w.disconnect()
+	if w.Instance != nil {
+		err = multierror.Append(err, w.Instance.Close()).ErrorOrNil()
 	}
-	return nil
+	if w.forwarder != nil {
+		w.forwarder.Close()
+	}
+	if w.ctx.Settings().FailOnDeprecation && w.sidecar != nil {
+		err = multierror.Append(err, w.checkDeprecation()).ErrorOrNil()
+	}
+	return
+}
+
+func (w *workload) checkDeprecation() error {
+	logs, err := w.sidecar.Logs()
+	if err != nil {
+		return fmt.Errorf("could not get sidecar logs to inspect for deprecation messages: %v", err)
+	}
+
+	info := fmt.Sprintf("pod: %s/%s", w.pod.Namespace, w.pod.Name)
+	return errors.FindDeprecatedMessagesInEnvoyLog(logs, info)
 }
 
 func (w *workload) PodName() string {
-	w.mutex.Lock()
-	n := w.pod.Name
-	w.mutex.Unlock()
-	return n
+	return w.pod.Name
 }
 
 func (w *workload) Address() string {
-	w.mutex.Lock()
-	ip := w.pod.Status.PodIP
-	w.mutex.Unlock()
-	return ip
+	return w.pod.Status.PodIP
 }
 
 func (w *workload) Sidecar() echo.Sidecar {
-	w.mutex.Lock()
-	s := w.sidecar
-	w.mutex.Unlock()
-	return s
+	return w.sidecar
 }
 
 func (w *workload) Logs() (string, error) {
@@ -139,82 +138,4 @@ func (w *workload) LogsOrFail(t test.Failer) string {
 		t.Fatal(err)
 	}
 	return logs
-}
-
-func isPodReady(pod kubeCore.Pod) bool {
-	return istioKube.CheckPodReady(&pod) == nil
-}
-
-func (w *workload) isConnected() bool {
-	return w.forwarder != nil
-}
-
-func (w *workload) connect(pod kubeCore.Pod) error {
-	// Create a forwarder to the command port of the app.
-	var forwarder istioKube.PortForwarder
-	if err := retry.UntilSuccess(func() error {
-		fw, err := w.cluster.NewPortForwarder(pod.Name, pod.Namespace, "", 0, int(w.grpcPort))
-		if err != nil {
-			return fmt.Errorf("failed creating new port forwarder for pod %s/%s: %v",
-				pod.Namespace, pod.Name, err)
-		}
-		if err = fw.Start(); err != nil {
-			fw.Close()
-			return fmt.Errorf("failed starting port forwarder for pod %s/%s: %v",
-				pod.Namespace, pod.Name, err)
-		}
-		forwarder = fw
-		return nil
-	}, retry.Delay(1*time.Second), retry.Timeout(10*time.Second)); err != nil {
-		return err
-	}
-
-	// Create a gRPC client to this workload.
-	c, err := client.New(forwarder.Address(), w.tls)
-	if err != nil {
-		forwarder.Close()
-		return fmt.Errorf("failed connecting to grpc client to pod %s/%s : %v",
-			pod.Namespace, pod.Name, err)
-	}
-
-	var s *sidecar
-	if w.hasSidecar {
-		if s, err = newSidecar(pod, w.cluster); err != nil {
-			return fmt.Errorf("failed creating sidecar for pod %s/%s: %v",
-				pod.Namespace, pod.Name, err)
-		}
-	}
-
-	// Store the resources.
-	w.forwarder = forwarder
-	w.Instance = c
-	w.sidecar = s
-
-	return nil
-}
-
-func (w *workload) disconnect() (err error) {
-	if w.Instance != nil {
-		err = multierror.Append(err, w.Instance.Close()).ErrorOrNil()
-		w.Instance = nil
-	}
-	if w.forwarder != nil {
-		w.forwarder.Close()
-		w.forwarder = nil
-	}
-	if w.ctx.Settings().FailOnDeprecation && w.sidecar != nil {
-		err = multierror.Append(err, w.checkDeprecation()).ErrorOrNil()
-		w.sidecar = nil
-	}
-	return err
-}
-
-func (w *workload) checkDeprecation() error {
-	logs, err := w.sidecar.Logs()
-	if err != nil {
-		return fmt.Errorf("could not get sidecar logs to inspect for deprecation messages: %v", err)
-	}
-
-	info := fmt.Sprintf("pod: %s/%s", w.pod.Namespace, w.pod.Name)
-	return errors.FindDeprecatedMessagesInEnvoyLog(logs, info)
 }
